+++ addons/discord_gd/classes/embed.gd
@@ -85,7 +85,7 @@ func set_timestamp(_timestamp = ''):
 func set_color(_color):
 	# RBG color
 	if typeof(_color) == TYPE_ARRAY:
-		color = (int(_color[0]) * 256 * 256) + (int(_color[1]) * 256) + int(_color[3])
+		color = (int(_color[0]) * 256 * 256) + (int(_color[1]) * 256) + int(_color[2])
 
 	# Hex color
 	elif typeof(_color) == TYPE_STRING and _color.begins_with('#'):