+++ CHANGELOG.md
@@ -4,8 +4,9 @@ Changelog
 This is a high-level changelog for each released versions of the plugin.
 For a more detailed list of past and incoming changes, see the commit history.
 
-Latest
+1.2.1
 ------
+- Fixed Embed.set_color() for an array of RGB values
 - Moved documentation to [Discord.gd](https://3ddelano.github.io/discord.gd)
 - Added `DiscordBot.guild_update` signal
 - Added `Helpers.iso2unix()` method